+++ CHANGELOG.md
@@ -3,6 +3,10 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+### unreleased
+### Fixed
+* Fixed an issue on Windows where the main window would flash white on startup.
+
 ### 0.9.0 - 2023-10-06
 ### Added
 * Added an option for manually setting the UI scale in preferences to fix scaling issues on some systems.